feat(payment_methods): fix clippy warnings
